+++ src/config.rs
@@ -106,6 +106,7 @@ pub struct Config {
     pub player_ignorelist: Option<RegexSet>,
     pub filter_script: Option<PathBuf>,
     pub use_track_start_timestamp: Option<bool>,
+    pub ignore_private_browsing: Option<bool>,
     pub listenbrainz: Option<Vec<ListenBrainzConfig>>,
 }
 
@@ -120,6 +121,7 @@ impl Config {
             player_ignorelist: Some(RegexSet::default()),
             filter_script: Some(PathBuf::new()),
             use_track_start_timestamp: Some(false),
+            ignore_private_browsing: Some(true),
             listenbrainz: Some(vec![ListenBrainzConfig {
                 url: Some(String::new()),
                 token: ListenBrainzToken::default(),
@@ -203,6 +205,10 @@ fn override_from_environment(config: &mut Config) -> Result<()> {
     replace_if_some(
         &mut config.use_track_start_timestamp,
         get_envvar("USE_TRACK_START_TIMESTAMP")?,
+    );
+    replace_if_some(
+        &mut config.ignore_private_browsing,
+        get_envvar("IGNORE_PRIVATE_BROWSING")?,
     );
 
     Ok(())
@@ -297,6 +303,7 @@ mod tests {
             std::env::set_var("MIN_PLAY_TIME", "30");
             std::env::set_var("FILTER_SCRIPT", "/tmp/filter.sh");
             std::env::set_var("USE_TRACK_START_TIMESTAMP", "true");
+            std::env::set_var("IGNORE_PRIVATE_BROWSING", "true");
         }
 
         override_from_environment(&mut config).unwrap();
@@ -321,6 +328,7 @@ mod tests {
             Some(Path::new("/tmp/filter.sh"))
         );
         assert_eq!(config.use_track_start_timestamp, Some(true));
+        assert_eq!(config.ignore_private_browsing, Some(true));
     }
 
     #[test]